+++ src/lisp/kernel/cmp/primitives.lisp
@@ -281,18 +281,14 @@
     
          ;; Primitives for Cleavir code
 
-         (primitive         "cm_check_index" :return-type (list :t* :t* :t*))
          (primitive         "cm_vref" :return-type (list :t* :t*))
          (primitive         "cm_vset" :return-type (list :t* :t* :t*))
          (primitive         "cc_ensure_valid_object" :t* (list :t*))
          (primitive         "cc_getPointer" :i8* (list :t*))
          (primitive-unwinds "cc_makeCell" :t* nil)
+         (primitive-unwinds "cc_checkBound" :size_t (list :t* :size_t :t*))
          (primitive         "cc_simpleBitVectorAref" :i8 (list :t* :size_t))
          (primitive         "cc_simpleBitVectorAset" :void (list :t* :size_t :i8))
-         (primitive-unwinds "cc_simpleFloatVectorAref" :single-float (list :t* :t*))
-         (primitive-unwinds "cc_simpleDoubleVectorAref" :double-float (list :t* :t*))
-         (primitive-unwinds "cc_simpleFloatVectorAset" :void (list :single-float :t* :t*))
-         (primitive-unwinds "cc_simpleDoubleVectorAset" :void (list :double-float :t* :t*))
          (primitive         "cc_initialize_gcroots_in_module" :void (list :gcroots-in-module* ; holder
                                                                       :t** ; root_address
                                                                       :size_t ; num_roots