+++ src/llvmo/intrinsics.cc
@@ -925,6 +925,20 @@ void cc_rewind_vaslist(vaslist va_args, void** register_save_areaP)
 }
 #endif
 
+size_t cc_checkBound(core::T_O* array, size_t bound, core::T_O* index) {
+  (void)array; // FIXME
+  core::T_sp tindex((gctools::Tagged)index);
+  if (tindex.fixnump()) {
+    core::Fixnum findex = tindex.unsafe_fixnum();
+    if ((0 <= findex) && (findex < bound)) return findex;
+    else {
+      // FIXME: use core::badIndexError instead
+      TYPE_ERROR(tindex, Cons_O::createList(cl::_sym_integer,
+                                            core::make_fixnum(0),
+                                            Cons_O::createList(core::make_fixnum(bound))));
+    }
+  } else TYPE_ERROR(tindex, cl::_sym_fixnum);
+}
 
 unsigned char cc_simpleBitVectorAref(core::T_O* tarray, size_t index) {
   core::SimpleBitVector_O* array = reinterpret_cast<core::SimpleBitVector_O*>(gctools::untag_general<core::T_O*>(tarray));
@@ -934,64 +948,6 @@ unsigned char cc_simpleBitVectorAref(core::T_O* tarray, size_t index) {
 void cc_simpleBitVectorAset(core::T_O* tarray, size_t index, unsigned char v) {
   core::SimpleBitVector_O* array = reinterpret_cast<core::SimpleBitVector_O*>(gctools::untag_general<core::T_O*>(tarray));
   (*array)[index] = v;
-}
-
-float cc_simpleFloatVectorAref(core::T_O* array, core::T_O* index) {
-  T_sp tarray((gctools::Tagged)array);
-  T_sp tindex((gctools::Tagged)index);
-  // type has already been checked by lisp compiler
-  SimpleVector_float_sp svarray = gc::As_unsafe<SimpleVector_float_sp>(tarray);
-  if (tindex.fixnump()) {
-    Fixnum findex = tindex.unsafe_fixnum();
-    if ((0 <= findex) && (findex < svarray->length()))
-      return (*svarray)[findex];
-    else TYPE_ERROR(tindex, Cons_O::createList(cl::_sym_integer,
-                                               core::make_fixnum(0),
-                                               Cons_O::createList(core::make_fixnum(svarray->length()))));
-  } else TYPE_ERROR(tindex, cl::_sym_fixnum);
-}
-
-double cc_simpleDoubleVectorAref(core::T_O* array, core::T_O* index) {
-  T_sp tarray((gctools::Tagged)array);
-  T_sp tindex((gctools::Tagged)index);
-  // type has already been checked by lisp compiler
-  SimpleVector_double_sp svarray = gc::As_unsafe<SimpleVector_double_sp>(tarray);
-  if (tindex.fixnump()) {
-    Fixnum findex = tindex.unsafe_fixnum();
-    if ((0 <= findex) && (findex < svarray->length()))
-      return (*svarray)[findex];
-    else TYPE_ERROR(tindex, Cons_O::createList(cl::_sym_integer,
-                                               core::make_fixnum(0),
-                                               Cons_O::createList(core::make_fixnum(svarray->length()))));
-  } else TYPE_ERROR(tindex, cl::_sym_fixnum);
-}
-
-void cc_simpleFloatVectorAset(float v, core::T_O* array, core::T_O* index) {
-  T_sp tarray((gctools::Tagged)array);
-  T_sp tindex((gctools::Tagged)index);
-  SimpleVector_float_sp svarray = gc::As_unsafe<SimpleVector_float_sp>(tarray);
-  if (tindex.fixnump()) {
-    Fixnum findex = tindex.unsafe_fixnum();
-    if ((0 <= findex) && (findex < svarray->length()))
-      (*svarray)[findex] = v;
-    else TYPE_ERROR(tindex, Cons_O::createList(cl::_sym_integer,
-                                               core::make_fixnum(0),
-                                               Cons_O::createList(core::make_fixnum(svarray->length()))));
-  } else TYPE_ERROR(tindex, cl::_sym_fixnum);
-}
-
-void cc_simpleDoubleVectorAset(double v, core::T_O* array, core::T_O* index) {
-  T_sp tarray((gctools::Tagged)array);
-  T_sp tindex((gctools::Tagged)index);
-  SimpleVector_double_sp svarray = gc::As_unsafe<SimpleVector_double_sp>(tarray);
-  if (tindex.fixnump()) {
-    Fixnum findex = tindex.unsafe_fixnum();
-    if ((0 <= findex) && (findex < svarray->length()))
-      (*svarray)[findex] = v;
-    else TYPE_ERROR(tindex, Cons_O::createList(cl::_sym_integer,
-                                               core::make_fixnum(0),
-                                               Cons_O::createList(core::make_fixnum(svarray->length()))));
-  } else TYPE_ERROR(tindex, cl::_sym_fixnum);
 }
 
 core::T_O** activationFrameReferenceFromClosure(core::T_O* closureRaw)